modify: 스크린 타임 지속 시간 duration의 타입을 long으로 변경
기존의 구현은 Double 타입 시간 단위를 취급했으나, 오차 누적 등의 이슈로 Long 타입 분 단위로 취급하기로 결정.

+++ src/main/java/dev/wetox/WetoxRESTful/screentime/CategoryScreenTimeResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CategoryScreenTimeResponse {
     private Category category;
-    private Double duration;
+    private Long duration;
 
     public CategoryScreenTimeResponse(CategoryScreenTime categoryScreenTime) {
         this.category = categoryScreenTime.getCategory();
